Обозначила точку входа и разделила на модули
